Implemented the syntax highlighting (semantic) using treesitter
This allows monaco to use the theme provided.
But there are still some bugs including with the new token provider
monaco doesn't know about block and indentation.

As on the rust side, for the future the parse and highlight method
should not be inside of OpenedFile as a good design pattern practices